exc111 concluido

+++ ex111/ex111.py
@@ -1,4 +1,4 @@
 from utilidadesCeV import moeda, dado
 
-valor = int(input("Digite um preco: "))
+valor = float(input("Digite um preco: "))
 moeda.resumo(valor, 10, 10)
